move BridgeDownloadParameterServiceTest and BlackDuckParametersServiceTest in their corresponding test package
